Pull out the model
