+++ micro-judge/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "micro-judge"
-version = "0.1.1"
+version = "0.1.2"
 authors = ["terkwood <38859656+Terkwood@users.noreply.github.com>"]
 edition = "2018"
 
@@ -11,3 +11,4 @@ r2d2_redis = "0.13.0"
 serde = "1.0.104"
 serde_derive = "1.0.104"
 bincode = "1.2.1"
+micro_model_moves = { path = "../micro-model/moves" }
