+++ micro-judge/src/lib.rs
@@ -1,4 +1,5 @@
 extern crate bincode;
+extern crate micro_model_moves;
 extern crate r2d2_redis;
 extern crate redis;
 extern crate serde;
