+++ micro-model/moves/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "micro_model_moves"
+version = "0.1.0"
+authors = ["terkwood <38859656+Terkwood@users.noreply.github.com>"]
+edition = "2018"
+
+[dependencies]
+bincode = "1.2.1"
+serde = "1.0.104"
+serde_derive = "1.0.104"
+uuid = { version = "0.8.1", features = ["serde", "v4"] }
